+++ package.json
@@ -29,7 +29,7 @@
     "zap": "node src-script/zap-start.js --logToStdout",
     "backend": "electron src-electron/main-process/electron-main.js",
     "build-spa": "quasar build",
-    "doc": "jsdoc src-electron src-shared -r -d ./generated-html/",
+    "doc": "jsdoc src-electron src-shared -r -u docs -d ./generated-html/ README.md",
     "apidoc": "jsdoc2md src-shared/**/*.js src-electron/**/*.js > docs/api.md",
     "format-code": "pretty-quick",
     "pretty-quick": "pretty-quick",
@@ -52,7 +52,7 @@
     "node-abi": "^2.18.0",
     "node-gyp": "^7.0.0",
     "node-pre-gyp": "^0.15.0",
-    "pino": "^6.4.1",
+    "pino": "^6.5.0",
     "promised-handlebars": "^2.0.1",
     "properties": "^1.2.1",
     "quasar": "^1.12.13",
