+++ src-electron/generator/helper-zap.js
@@ -15,10 +15,26 @@
  *    limitations under the License.
  */
 
+/**
+ * This module contains the API for templating. For more detailed instructions, read {@tutorial template-tutorial}
+ *
+ * @module Templating API: toplevel utility helpers
+ */
+
+/**
+ * Produces the top-of-the-file header for a C file.
+ *
+ * @returns The header content
+ */
 function zap_header() {
   return `// This file is generated by ZCL Advanced Platform generator. Please don't edit manually.`
 }
 
+/**
+ * Simple helper that produces an approved size of identation.
+ *
+ * @returns whitespace that is the identation.
+ */
 function ident() {
   return '  '
 }
