+++ src-electron/generator/helper-zcl.js
@@ -16,36 +16,14 @@
  */
 
 const queryZcl = require('../db/query-zcl.js')
-const queryPackage = require('../db/query-package.js')
 const dbEnum = require('../db/db-enum.js')
 const templateUtil = require('./template-util.js')
 
 /**
- * Returns the promise that resolves with the ZCL properties package id.
+ * This module contains the API for templating. For more detailed instructions, read {@tutorial template-tutorial}
  *
- * @param {*} context
- * @returns promise that resolves with the package id.
+ * @module Templating API: static zcl helpers
  */
-function ensurePackageId(context) {
-  if ('packageId' in context.global) {
-    return Promise.resolve(context.global.packageId)
-  } else {
-    return queryPackage
-      .getSessionPackagesByType(
-        context.global.db,
-        context.global.sessionId,
-        dbEnum.packageType.zclProperties
-      )
-      .then((pkgs) => {
-        if (pkgs.length == 0) {
-          return null
-        } else {
-          context.global.packageId = pkgs[0].id
-          return pkgs[0].id
-        }
-      })
-  }
-}
 
 /**
  * Block helper iterating over all enums.
@@ -54,7 +32,8 @@ function ensurePackageId(context) {
  * @returns Promise of content.
  */
 function zcl_enums(options) {
-  return ensurePackageId(this)
+  return templateUtil
+    .ensurePackageId(this)
     .then((packageId) => queryZcl.selectAllEnums(this.global.db, packageId))
     .then((ens) => templateUtil.collectBlocks(ens, options.fn, this))
 }
@@ -77,7 +56,8 @@ function zcl_enum_items(options) {
  * @returns Promise of content.
  */
 function zcl_structs(options) {
-  return ensurePackageId(this)
+  return templateUtil
+    .ensurePackageId(this)
     .then((packageId) => queryZcl.selectAllStructs(this.global.db, packageId))
     .then((st) => templateUtil.collectBlocks(st, options.fn, this))
 }
@@ -101,7 +81,8 @@ function zcl_struct_items(options) {
  * @returns Promise of content.
  */
 function zcl_clusters(options) {
-  return ensurePackageId(this)
+  return templateUtil
+    .ensurePackageId(this)
     .then((packageId) => queryZcl.selectAllClusters(this.global.db, packageId))
     .then((cl) => templateUtil.collectBlocks(cl, options.fn, this))
 }
@@ -116,7 +97,8 @@ function zcl_clusters(options) {
  * @returns Promise of content.
  */
 function zcl_commands(options) {
-  return ensurePackageId(this)
+  return templateUtil
+    .ensurePackageId(this)
     .then((packageId) => {
       if ('id' in this) {
         // We're functioning inside a nested context with an id, so we will only query for this cluster.
@@ -132,10 +114,19 @@ function zcl_commands(options) {
     .then((cmds) => templateUtil.collectBlocks(cmds, options.fn, this))
 }
 
+/**
+ * Iterator over the attributes. If it is used at toplevel, if iterates over all the attributes
+ * in the database. If used within zcl_cluster context, it iterates over all the attributes
+ * that belong to that cluster.
+ *
+ * @param {*} options
+ * @returns Promise of attribute iteration.
+ */
 function zcl_attributes(options) {
   // If used at the toplevel, 'this' is the toplevel context object.
   // when used at the cluster level, 'this' is a cluster
-  return ensurePackageId(this)
+  return templateUtil
+    .ensurePackageId(this)
     .then((packageId) => {
       if ('id' in this) {
         // We're functioning inside a nested context with an id, so we will only query for this cluster.
@@ -151,10 +142,19 @@ function zcl_attributes(options) {
     .then((atts) => templateUtil.collectBlocks(atts, options.fn, this))
 }
 
+/**
+ * Iterator over the client attributes. If it is used at toplevel, if iterates over all the client attributes
+ * in the database. If used within zcl_cluster context, it iterates over all the client attributes
+ * that belong to that cluster.
+ *
+ * @param {*} options
+ * @returns Promise of attribute iteration.
+ */
 function zcl_attributes_client(options) {
   // If used at the toplevel, 'this' is the toplevel context object.
   // when used at the cluster level, 'this' is a cluster
-  return ensurePackageId(this)
+  return templateUtil
+    .ensurePackageId(this)
     .then((packageId) => {
       if ('id' in this) {
         return queryZcl.selectAttributesByClusterIdAndSide(
@@ -174,10 +174,19 @@ function zcl_attributes_client(options) {
     .then((atts) => templateUtil.collectBlocks(atts, options.fn, this))
 }
 
+/**
+ * Iterator over the server attributes. If it is used at toplevel, if iterates over all the server attributes
+ * in the database. If used within zcl_cluster context, it iterates over all the server attributes
+ * that belong to that cluster.
+ *
+ * @param {*} options
+ * @returns Promise of attribute iteration.
+ */
 function zcl_attributes_server(options) {
   // If used at the toplevel, 'this' is the toplevel context object.
   // when used at the cluster level, 'this' is a cluster
-  return ensurePackageId(this)
+  return templateUtil
+    .ensurePackageId(this)
     .then((packageId) => {
       if ('id' in this) {
         // We're functioning inside a nested context with an id, so we will only query for this cluster.
@@ -205,7 +214,8 @@ function zcl_attributes_server(options) {
  * @returns Promise of content.
  */
 function zcl_atomics(options) {
-  return ensurePackageId(this)
+  return templateUtil
+    .ensurePackageId(this)
     .then((packageId) => queryZcl.selectAllAtomics(this.global.db, packageId))
     .then((ats) => templateUtil.collectBlocks(ats, options.fn, this))
 }
