+++ test/gen-template/zap-config.handlebars
@@ -5,5 +5,11 @@
 
 {{#user_clusters}}
 // Cluster: {{code}} / {{name}} ({{side}}) / {{mfgCode}}
+{{#user_cluster_attributes}}
+// - attribute: {{code}} / {{name}}
+{{/user_cluster_attributes}}
+{{#user_cluster_commands}}
+// - command: {{code}} / {{name}}
+{{/user_cluster_commands}}
 {{/user_clusters}}
 {{/user_endpoint_types}}
